feat: script runner
