+++ plugin-script/build.gradle
@@ -17,4 +17,6 @@ dependencies {
         exclude group: 'com.github.docker-java', module: 'docker-java-transport-jersey'
     }
     api 'com.github.docker-java:docker-java-transport-zerodep:3.3.5'
+
+    testImplementation group: "io.kestra", name: "core", version: kestraVersion, classifier: 'tests'
 }
