--- plugin-script/src/main/java/io/kestra/plugin/scripts/exec/scripts/runners/RunnerResult.java
@@ -1,11 +1,0 @@
-package io.kestra.plugin.scripts.exec.scripts.runners;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor
-@Getter
-public class RunnerResult {
-    private int exitCode;
-    private AbstractLogConsumer logConsumer;
-}
